Front end has a dashboard, products page, and inventory page. All fucntion properly.

+++ src/main/java/com/abrams/projectone/inventory/InventoryService.java
@@ -43,7 +43,7 @@ public class InventoryService {
 
         // 3. CAPACITY CHECK (Existing logic)
         Integer maxCapacity = warehouse.getMaxCapacity();
-        Integer currentLoad = inventoryRepository.getTotalQuantityInWarehouse(warehouseId);
+        Long currentLoad = inventoryRepository.getTotalQuantityInWarehouse(warehouseId);
         
         if (currentLoad + inventory.getQuantity() > maxCapacity) {
              throw new RuntimeException("Warehouse capacity exceeded!");
